refactor: move dumbo, mmds and logger to vmm (7/11)
Rename vmm::config::logger to avoid having conflicts with
vmm::logger crate.

Note: This is part of commits having the same title.
