+++ src/vmm/src/vmm_config/mod.rs
@@ -23,7 +23,7 @@ pub mod entropy;
 /// Wrapper over the microVM general information attached to the microVM.
 pub mod instance_info;
 /// Wrapper for configuring the logger.
-pub mod logger;
+pub mod logger_config;
 /// Wrapper for configuring the memory and CPU of the microVM.
 pub mod machine_config;
 /// Wrapper for configuring the metrics.
